+++ test/unit/image.js
@@ -195,14 +195,14 @@
     var done = assert.async();
     createImageObject(function(image) {
       assert.ok(typeof image.toObject === 'function');
-      var filter = new fabric.Image.filters.Resize({resizeType: 'bilinear', scaleX: 0.3, scaleY: 0.3});
+      var filter = new fabric.filters.Resize({resizeType: 'bilinear', scaleX: 0.3, scaleY: 0.3});
       image.resizeFilter = filter;
-      assert.ok(image.resizeFilter instanceof fabric.Image.filters.Resize, 'should inherit from fabric.Image.filters.Resize');
+      assert.ok(image.resizeFilter instanceof fabric.filters.Resize, 'should inherit from fabric.filters.Resize');
       var toObject = image.toObject();
       assert.deepEqual(toObject.resizeFilter, filter.toObject(), 'the filter is in object form now');
       fabric.Image.fromObject(toObject).then(function(imageFromObject) {
         var filterFromObj = imageFromObject.resizeFilter;
-        assert.ok(filterFromObj instanceof fabric.Image.filters.Resize, 'should inherit from fabric.Image.filters.Resize');
+        assert.ok(filterFromObj instanceof fabric.filters.Resize, 'should inherit from fabric.filters.Resize');
         assert.deepEqual(filterFromObj, filter,  'the filter has been restored');
         assert.equal(filterFromObj.scaleX, 0.3);
         assert.equal(filterFromObj.scaleY, 0.3);
@@ -215,9 +215,9 @@
   QUnit.test('toObject with normal filter and resize filter', function(assert) {
     var done = assert.async();
     createImageObject(function(image) {
-      var filter = new fabric.Image.filters.Resize({resizeType: 'bilinear' });
+      var filter = new fabric.filters.Resize({resizeType: 'bilinear' });
       image.resizeFilter = filter;
-      var filterBg = new fabric.Image.filters.Brightness({ brightness: 0.8 });
+      var filterBg = new fabric.filters.Brightness({ brightness: 0.8 });
       image.filters = [filterBg];
       image.scaleX = 0.3;
       image.scaleY = 0.3;
@@ -227,8 +227,8 @@
       fabric.Image.fromObject(toObject).then(function(imageFromObject) {
         var filterFromObj = imageFromObject.resizeFilter;
         var brightnessFromObj = imageFromObject.filters[0];
-        assert.ok(filterFromObj instanceof fabric.Image.filters.Resize, 'should inherit from fabric.Image.filters.Resize');
-        assert.ok(brightnessFromObj instanceof fabric.Image.filters.Brightness, 'should inherit from fabric.Image.filters.Resize');
+        assert.ok(filterFromObj instanceof fabric.filters.Resize, 'should inherit from fabric.filters.Resize');
+        assert.ok(brightnessFromObj instanceof fabric.filters.Brightness, 'should inherit from fabric.filters.Resize');
         done();
       });
     });
@@ -238,10 +238,10 @@
     var done = assert.async();
     createImageObject(function(image) {
       assert.ok(typeof image.toObject === 'function');
-      var filter = new fabric.Image.filters.Resize({resizeType: 'bilinear', scaleX: 0.2, scaleY: 0.2});
+      var filter = new fabric.filters.Resize({resizeType: 'bilinear', scaleX: 0.2, scaleY: 0.2});
       image.filters.push(filter);
       var width = image.width, height = image.height;
-      assert.ok(image.filters[0] instanceof fabric.Image.filters.Resize, 'should inherit from fabric.Image.filters.Resize');
+      assert.ok(image.filters[0] instanceof fabric.filters.Resize, 'should inherit from fabric.filters.Resize');
       image.applyFilters();
       assert.equal(image.width, Math.floor(width), 'width is not changed');
       assert.equal(image.height, Math.floor(height), 'height is not changed');
@@ -253,7 +253,7 @@
       assert.equal(toObject.height, height, 'height is stored as before filters');
       fabric.Image.fromObject(toObject).then(function(_imageFromObject) {
         var filterFromObj = _imageFromObject.filters[0];
-        assert.ok(filterFromObj instanceof fabric.Image.filters.Resize, 'should inherit from fabric.Image.filters.Resize');
+        assert.ok(filterFromObj instanceof fabric.filters.Resize, 'should inherit from fabric.filters.Resize');
         assert.equal(filterFromObj.scaleY, 0.2);
         assert.equal(filterFromObj.scaleX, 0.2);
         done();
@@ -486,8 +486,8 @@
       assert.ok(instance.clipPath instanceof fabric.Rect);
       assert.ok(Array.isArray(instance.filters), 'should enliven filters');
       assert.equal(instance.filters.length, 1, 'should enliven filters');
-      assert.ok(instance.filters[0] instanceof fabric.Image.filters.Brightness, 'should enliven filters');
-      assert.ok(instance.resizeFilter instanceof fabric.Image.filters.Resize, 'should enliven resizeFilter');
+      assert.ok(instance.filters[0] instanceof fabric.filters.Brightness, 'should enliven filters');
+      assert.ok(instance.resizeFilter instanceof fabric.filters.Resize, 'should enliven resizeFilter');
       done();
     });
   });
@@ -805,7 +805,7 @@
     createImageObject(function(image) {
       var run = false;
       image.dirty = false;
-      var filter = new fabric.Image.filters.Brightness();
+      var filter = new fabric.filters.Brightness();
       image.filters = [filter];
       filter.isNeutralState = function() {
         run = true;
@@ -831,7 +831,7 @@
   QUnit.test('apply filters reset _element and _filteredEl', function(assert) {
     var done = assert.async();
     createImageObject(function(image) {
-      var contrast = new fabric.Image.filters.Contrast({ contrast: 0.5 });
+      var contrast = new fabric.filters.Contrast({ contrast: 0.5 });
       image.applyFilters();
       var element = image._element;
       var filtered = image._filteredEl;
@@ -847,8 +847,8 @@
   QUnit.test('apply filters and resize filter', function(assert) {
     var done = assert.async();
     createImageObject(function(image) {
-      var contrast = new fabric.Image.filters.Contrast({ contrast: 0.5 });
-      var resizeFilter = new fabric.Image.filters.Resize();
+      var contrast = new fabric.filters.Contrast({ contrast: 0.5 });
+      var resizeFilter = new fabric.filters.Resize();
       image.filters = [contrast];
       image.resizeFilter = resizeFilter;
       var element = image._element;